# data/binding: use atomics