--- a/data/binding/listbinding_test.go
+++ b/data/binding/listbinding_test.go
@@ -1,6 +1,7 @@
 package binding
 
 import (
+	"sync"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -12,14 +13,14 @@ type simpleList struct {
 
 func TestListBase_AddListener(t *testing.T) {
 	data := &simpleList{}
-	assert.Equal(t, 0, len(data.listeners))
+	assert.Equal(t, 0, syncMapLen(&data.listeners))
 
 	called := false
 	fn := NewDataListener(func() {
 		called = true
 	})
 	data.AddListener(fn)
-	assert.Equal(t, 1, len(data.listeners))
+	assert.Equal(t, 1, syncMapLen(&data.listeners))
 
 	data.trigger()
 	waitForItems()
@@ -56,11 +57,13 @@ func TestListBase_RemoveListener(t *testing.T) {
 		called = true
 	})
 	data := &simpleList{}
-	data.listeners = []DataListener{fn}
+	m := sync.Map{}
+	m.Store(fn, true)
+	data.listeners = m
 
-	assert.Equal(t, 1, len(data.listeners))
+	assert.Equal(t, 1, syncMapLen(&data.listeners))
 	data.RemoveListener(fn)
-	assert.Equal(t, 0, len(data.listeners))
+	assert.Equal(t, 0, syncMapLen(&data.listeners))
 
 	data.trigger()
 	waitForItems()