--- a/data/binding/pref_helper.go
+++ b/data/binding/pref_helper.go
@@ -11,72 +11,54 @@ type preferenceItem interface {
 }
 
 type preferenceBindings struct {
-	lock  sync.RWMutex
-	items map[string]preferenceItem
+	items sync.Map // map[string]preferenceItem
 }
 
 func (b *preferenceBindings) getItem(key string) preferenceItem {
-	b.lock.RLock()
-	item := b.items[key]
-	b.lock.RUnlock()
-	return item
+	val, loaded := b.items.Load(key)
+	if !loaded {
+		return nil
+	}
+	return val.(preferenceItem)
 }
 
 func (b *preferenceBindings) list() []preferenceItem {
-	b.lock.RLock()
-	allItems := b.items
-	b.lock.RUnlock()
-	ret := make([]preferenceItem, 0, len(allItems))
-	for _, i := range allItems {
-		ret = append(ret, i)
-	}
+	ret := []preferenceItem{}
+	b.items.Range(func(_, val interface{}) bool {
+		ret = append(ret, val.(preferenceItem))
+		return true
+	})
 	return ret
 }
 
 func (b *preferenceBindings) setItem(key string, item preferenceItem) {
-	b.lock.Lock()
-	b.items[key] = item
-	b.lock.Unlock()
+	b.items.Store(key, item)
 }
 
 type preferencesMap struct {
-	lock  sync.RWMutex
-	prefs map[fyne.Preferences]*preferenceBindings
+	prefs sync.Map // map[fyne.Preferences]*preferenceBindings
 }
 
 func newPreferencesMap() *preferencesMap {
-	return &preferencesMap{
-		prefs: make(map[fyne.Preferences]*preferenceBindings),
-	}
+	return &preferencesMap{}
 }
 
 func (m *preferencesMap) ensurePreferencesAttached(p fyne.Preferences) *preferenceBindings {
-	m.lock.RLock()
-	binds := m.prefs[p]
-	m.lock.RUnlock()
-
-	if binds != nil {
-		return binds
+	binds, loaded := m.prefs.LoadOrStore(p, &preferenceBindings{})
+	if loaded {
+		return binds.(*preferenceBindings)
 	}
 
-	m.lock.Lock()
-	m.prefs[p] = &preferenceBindings{
-		items: make(map[string]preferenceItem),
-	}
-	binds = m.prefs[p]
-	m.lock.Unlock()
-
-	p.AddChangeListener(func() {
-		m.preferencesChanged(p)
-	})
-	return binds
+	p.AddChangeListener(func() { m.preferencesChanged(p) })
+	return binds.(*preferenceBindings)
 }
 
 func (m *preferencesMap) getBindings(p fyne.Preferences) *preferenceBindings {
-	m.lock.RLock()
-	binds := m.prefs[p]
-	m.lock.RUnlock()
-	return binds
+	binds, loaded := m.prefs.Load(p)
+	if !loaded {
+		return nil
+	}
+	return binds.(*preferenceBindings)
 }
 
 func (m *preferencesMap) preferencesChanged(p fyne.Preferences) {