### Description:
Use `sync/atomic.*` and `sync.Map` to optimize the following two cases:

1. atomic access on an individual variable
2. read-heavy cases, such as preference maps

Existing benchmarks:

```
name             old time/op  new time/op  delta
BoolToString-8   1.12µs ± 2%  1.10µs ± 2%    ~     (p=0.095 n=5+5)
FloatToString-8  1.42µs ± 0%  1.37µs ± 0%  -3.08%  (p=0.016 n=4+5)
IntToString-8    1.13µs ± 0%  1.12µs ± 2%    ~     (p=0.159 n=4+5)
StringToBool-8   1.46µs ± 2%  1.39µs ± 0%  -5.04%  (p=0.008 n=5+5)
StringToFloat-8  1.60µs ± 1%  1.51µs ± 0%  -5.62%  (p=0.000 n=5+4)
StringToInt-8    1.47µs ± 2%  1.41µs ± 0%  -4.17%  (p=0.008 n=5+5)
```

The actual performance gain might be even higher, but concerning new benchmarks would be another story.

Fixes #2359

### Checklist:
<!-- Please tick these as appropriate using [x] -->

- [x] Tests included.
- [x] Lint and formatter run with no errors.
- [x] Tests all pass.