@Bluebugs 2021-11-23
I think that writing meaningful benchmark for locking mechanism is quite hard. In this case, a RWLock would take more memory and might trigger in some case an overhead that the Atomic would not. It also feel safer in term of code as you have to use the Atomic API, you can't forget it when accessing the value. So even without the benchmark, I think this is a meaningful improvement.